Save locations to database
Insert the location updates into the database. Use string
representations for the floating point parameters for simplicity's sake
(although perhaps not brevity's).

Pass the connection into the callback via the user data parameter; in
order to use it, though, I had to figure out that there was a second
parameter in between - I can't find where this is documented!